erika_cloud/erika_cloud_django/pyproject.toml
[tool.poetry]
name = "erika-cloud-django"
version = "0.1.0"
description = "Cloud Backend for Erika Typrwriters based on Django"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "^1.1.0"
dj-database-url = "^2.1.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
